Waiting for internet connection

+++ scripts/rasp-setup.sh
@@ -1,21 +1,44 @@
 #!/bin/bash
 
 BRIDGE_FILE=/etc/systemd/network/bridge.network
+DOCKER_SRC_FILE=/etc/apt/sources.list.d/docker.list
+
+function_test_internet_connection() {
+    if nc -zw1 google.com 443; then
+        echo "The network is up!"
+    else
+        echo "The network is down..."
+    fi
+}
+
+function_check_pkg_installed() {
+    pkgs=("$@")
+    for i in "${pkgs[@]}"
+    do
+        dpkg -s $i &> /dev/null
+        if [ $? -eq 0 ]; then
+            echo "Package is installed!"
+        else
+            sudo apt-get --assume-yes --no-install-recommends install $i
+        fi
+    done
+}
 
 echo "========== Upgrading pkgs =========="
+sudo apt-get --assume-yes update && sudo apt-get --assume-yes upgrade
 
-sudo apt update -y && sudo apt upgrade -y
+curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
 
-echo "========== Removing file =========="
-sudo apt-get remove -y docker \
-                       docker-ce \
-                       docker-engine \
-                       docker.io \
-                       containerd \
-                       runc
+echo "========== Removing old unwanted docker-related pkgs =========="
+sudo apt-get --assume-yes remove docker \
+                                 docker-ce \
+                                 docker-engine \
+                                 docker.io \
+                                 containerd \
+                                 runc
 
 echo "========== Creating Bridge file =========="
-if [ ! -f $BRIDGE_FILE ]; then
+if [ ! -f "$BRIDGE_FILE" ]; then
     sudo tee -a $BRIDGE_FILE > /dev/null << EOF
 [Network]
 
@@ -27,41 +50,31 @@ else
     echo $BRIDGE_FILE && cat $BRIDGE_FILE
 fi
 
-echo "========== Setting Bash color =========="
-echo -e "
-function git_branch_name() {
-    git branch 2>/dev/null | grep -e '^*' | sed -E 's/^\* (.+)$/(\1)/'
-}
-
-function show_colored_git_branch_in_prompt() {\n
-    PS1='\[\033[01;32m\]\u@\h:\[\033[01;34m\]\w\[\033[31m\]\$(git_branch_name)\[\033[m\]$'
-}
-
-show_colored_git_branch_in_prompt
-" >> ~/.bashrc
-source ~/.bashrc
-
 echo "========== Restarting Network service =========="
 sudo systemctl restart systemd-networkd.service
 
+until function_test_internet_connection; do
+  echo "Network is still unavailable, waiting..."
+  sleep 2
+done
+
 echo "========== Installing required Docker dependencies =========="
-sudo apt-get install --no-install-recommends \
-                        apt-transport-https \
-                        ca-certificates \
-                        curl \
-                        gnupg2 \
-                        software-properties-common
+packages=( "apt-transport-https" "ca-certificates" "curl" "gnupg2" "software-properties-common" )
+function_check_pkg_installed "${packages[@]}"
 
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
-
-sudo tee -a /etc/apt/sources.list.d/docker.list > /dev/null << EOF
+if [ ! -f "$DOCKER_SRC_FILE" ]; then
+   sudo tee -a /etc/apt/sources.list.d/docker.list > /dev/null << EOF
 deb [arch=armhf] https://download.docker.com/linux/debian buster stable
 EOF
-
-sudo apt-get update
+   cat $DOCKER_SRC_FILE
+else
+   echo -e "Docker source file already exists...\n"
+   echo $DOCKER_SRC_FILE && cat $DOCKER_SRC_FILE
+fi
+sudo apt-get --assume-yes update
 
 echo "========== Installing Docker =========="
-sudo apt-get install --no-install-recommends docker-ce
+function_check_pkg_installed docker-ce
 
 echo "========== Adding pi to docker group =========="
 sudo usermod -aG docker pi
@@ -69,21 +82,23 @@ sudo usermod -aG docker pi
 echo "========== Installing Docker Compose =========="
 sudo pip install docker-compose
 
+sudo apt-get --assume-yes autoremove
+
 echo "========== SUCESSFULLY INSTALLED ALL NEEDED PKGs =========="
 
-echo "========== Cloning SMI-Slave project =========="
+echo "========== Setting SMI-Slave project =========="
+
 mkdir -p /smi
 git clone https://gitlab.com/lappis-unb/projects/SMI/smi-slave.git /smi/slave
 
-echo "*Successfully cloned SMI-Slave into /smi/slave"
+function_ask_reboot() {
+    read -p "Do you want to reboot now? (y|n)    " INPUT
 
-read -p "Do you want to reboot now? Yes | No" INPUT
-if [ $INPUT == 'Yes' ]; then
-    echo "========== Rebooting to apply changes =========="
-    sudo reboot
-if [ $INPUT == 'No' ]; then
-    echo "You must reboot to apply changes!"
-else
-    echo "Type 'Yes' or 'No'"
-fi
-fi
+    case "$INPUT" in 
+    y|Y ) echo "========== Rebooting to apply changes ==========" && sudo reboot;;
+    n|N ) echo "You must reboot to apply changes!";;
+    * ) echo "Type 'Y' or 'N'" && function_ask_reboot;;
+    esac
+}
+
+function_ask_reboot
